Create utils/folders module
